Added blueprinting and basic error pages
