+++ app/case/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+case_bp = Blueprint('case', __name__)
+
+from app.case import routes
