+++ app/errors/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+errors_bp = Blueprint('errors', __name__)
+
+from app.errors import handlers
