+++ app/templates/errors/404.html
@@ -1,0 +1,14 @@
+{%- extends 'base.html' -%}
+
+{%- set pageConfig = {
+    "title": "Service name",
+    "footer": {
+    }
+} -%}
+
+{%- block main -%}
+    <h1>Page not found</h1>
+    <p>If you entered a web address, check it is correct.</p>
+    <p>If you pasted the web address, check you copied the whole address.</p>
+    <p>If the web address is correct or you selected a link or button, <a href="#0">contact us</a> for more help.</p>
+{%- endblock -%}
