+++ app/templates/home.html
@@ -15,7 +15,7 @@
                     "id": 'title1',
                     "textId": 'text1',
                     "title": 'Find case via postcode',
-                    "url": url_for('postcode_input'),
+                    "url": url_for('address_search.postcode_input'),
                     "text": 'Find the callers case via postcode search'
                 }) }}
             </div>
@@ -24,7 +24,7 @@
                     "id": 'title2',
                     "textId": 'text2',
                     "title": 'Find case via address text search',
-                    "url": url_for('address_input'),
+                    "url": url_for('address_search.address_input'),
                     "text": 'Find the callers case via address text search'
                 }) }}
             </div>
